Move signac wrappers to own subpackage
